Ensure safe manipulation of bundled files buffer
The `Watch#start()` method should not clear the buffer of bundled files, because we don't control when it's invoked. If it is invoked a second time, and `Watch#capture` wasn't called in the meantime, this will lead to unwatching all files and a Karma process that appears to be stuck becuase it no longer reacts to file changes. Clearing the buffer before re-populating it is the safe way to handle this.