Fix stale bundles with shared dependecies
When we have two entries E1 and E2 and both depend on module A, we
would previously only update the first entry (E1). We never refreshed
E2 making it contain old code.

So we need to collect all entry files which include the changed file
somwhere in their dependency tree.